Organize iOS Drop-in lifecycle tests
